test(settlement): move new nonce lock tests into per-module files
Follow the agglayer-storage convention of keeping unit tests in
per-module test files instead of growing the inline test modules.
The tests stay unit tests (they construct types with module-private
fields), so they move to child-module files rather than tests/.
